EPS32_Adapter: Moved part of code into source file
